--HG--
extra : convert_revision : e802c800a478c297d3aa780a9ea3c6701453d91d

+++ base/stats/events.cc
@@ -134,7 +134,7 @@ InsertEvent::insert(const string &stat)
     first = false;
 
     size += sprintf(query + size, "(%u,%u,%llu)",
-                    event, run, curTick);
+                    event, run, (unsigned long long)curTick);
 }
 
 void
